# openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/ideas":{"get":{"summary":"Get Ideas","operationId":"get_ideas_ideas_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdeaList"}}}}}},"post":{"summary":"Add Idea","operationId":"add_idea_ideas_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdeaIn"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdeaList"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ideas/{idea_id}":{"get":{"summary":"Get Idea","operationId":"get_idea_ideas__idea_id__get","parameters":[{"required":true,"schema":{"title":"Idea Id"},"name":"idea_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Idea"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Update Idea","operationId":"update_idea_ideas__idea_id__put","parameters":[{"required":true,"schema":{"title":"Idea Id"},"name":"idea_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdeaIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Idea"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Idea","operationId":"delete_idea_ideas__idea_id__delete","parameters":[{"required":true,"schema":{"title":"Idea Id"},"name":"idea_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdeaList"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"summary":"Overwrite Idea","operationId":"overwrite_idea_ideas__idea_id__patch","parameters":[{"required":true,"schema":{"title":"Idea Id"},"name":"idea_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdeaIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Idea"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Idea":{"title":"Idea","required":["title","desc"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"title":{"title":"Title","type":"string"},"desc":{"title":"Desc","type":"string"},"date_created":{"title":"Date Created","type":"string","format":"date"},"updated":{"title":"Updated","type":"string","format":"date-time"}}},"IdeaIn":{"title":"IdeaIn","required":["title","desc"],"type":"object","properties":{"title":{"title":"Title","type":"string"},"desc":{"title":"Desc","type":"string"}}},"IdeaList":{"title":"IdeaList","required":["ideas"],"type":"object","properties":{"ideas":{"title":"Ideas","type":"object","additionalProperties":{"$ref":"#/components/schemas/Idea"}}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}