# src/openapi_server/models/http_validation_error.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401
from openapi_server.models.validation_error import ValidationError


class HTTPValidationError(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    HTTPValidationError - a model defined in OpenAPI

        detail: The detail of this HTTPValidationError [Optional].
    """

    detail: Optional[List[ValidationError]] = Field(alias="detail", default=None)

HTTPValidationError.update_forward_refs()