# src/openapi_server/models/idea_in.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401


class IdeaIn(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    IdeaIn - a model defined in OpenAPI

        title: The title of this IdeaIn.
        desc: The desc of this IdeaIn.
    """

    title: str = Field(alias="title")
    desc: str = Field(alias="desc")

IdeaIn.update_forward_refs()