# src/openapi_server/models/idea_list.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401
from openapi_server.models.idea import Idea


class IdeaList(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    IdeaList - a model defined in OpenAPI

        ideas: The ideas of this IdeaList.
    """

    ideas: Dict[str, Idea] = Field(alias="ideas")

IdeaList.update_forward_refs()